Address PR comments. 1) remove blocking call; 2) scope out encrypt / decrypt as a helper class
